+++ instrumentor/src/index.js
@@ -20,6 +20,14 @@ const buildStartupWrapper = template(`{
   ORIGINAL_FUNCTION_CALL
 }`);
 
+const buildInstrumentationHoc = template(`
+  const Heap = require('@heap/react-native-heap').default || {
+    HOC_IDENTIFIER: (Component) => Component,
+  };
+
+  const COMPONENT_ID = HOC_CALL_EXPRESSION;
+`);
+
 const identifierVisitor = {
   Identifier(path) {
     if (path.node.name === 'Touchable') {
@@ -295,29 +303,41 @@ const instrumentStartup = path => {
   }
 };
 
-const pressabilityWrappingTemplate = template(`
-  const Heap = require('@heap/react-native-heap').default;
-  CONFIG_IDENTIFIER = Heap.wrapPressabilityConfig(CONFIG_IDENTIFIER) || CONFIG_IDENTIFIER;
-`);
+const TOUCHABLE_COMPONENTS = [
+  'TouchableOpacity',
+  'TouchableNativeFeedback',
+  'TouchableWithoutFeedback',
+  'TouchableHighlight',
+];
 
-// Instruments 'Touchable's in RN 0.62+ by passing the 'Pressability' config to the Heap library to wrap the config.  See
-// https://github.com/facebook/react-native/blob/a5151c2b5f6f03896eb7d9df873c5f61a706f055/Libraries/Components/Touchable/TouchableOpacity.js#L139
-// and
-// https://github.com/facebook/react-native/blob/a5151c2b5f6f03896eb7d9df873c5f61a706f055/Libraries/Pressability/Pressability.js#L405-L407.
-const pressabilityInstrumentationVisitor = {
-  ClassMethod(path) {
-    const { node } = path;
+const instrumentTouchableHoc = path => {
+  if (!TOUCHABLE_COMPONENTS.includes(path.node.id.name)) {
+    return;
+  }
 
-    if (node.key.name !== 'constructor') {
-      return;
-    }
+  // 'path.node' represents a class *declaration*, so we need to convert 'path.node' to a class *expression* before we can pass it as an
+  // argument to our HOC function.
+  const equivalentExpression = t.classExpression(
+    path.node.id,
+    path.node.superClass,
+    path.node.body,
+    path.node.decorators || []
+  );
 
-    const configWrapExpression = pressabilityWrappingTemplate({
-      CONFIG_IDENTIFIER: node.params[0],
-    });
+  const hocIdentifier = t.identifier('withHeapTouchableAutocapture');
 
-    path.get('body').unshiftContainer('body', configWrapExpression);
-  },
+  const autotrackExpression = t.callExpression(
+    t.memberExpression(t.identifier('Heap'), hocIdentifier),
+    [equivalentExpression]
+  );
+
+  const replacement = buildInstrumentationHoc({
+    COMPONENT_ID: path.node.id,
+    HOC_IDENTIFIER: hocIdentifier,
+    HOC_CALL_EXPRESSION: autotrackExpression,
+  });
+
+  path.replaceWithMultiple(replacement);
 };
 
 function transform(babel) {
@@ -333,9 +353,7 @@ function transform(babel) {
         instrumentSwitchComponent(path);
       },
       ClassDeclaration(path) {
-        if (path.node.id.name === 'Pressability') {
-          path.traverse(pressabilityInstrumentationVisitor);
-        }
+        instrumentTouchableHoc(path);
       },
     },
   };