+++ instrumentor/test/fixtures/is-pressability/code.js
@@ -1,7 +1,0 @@
-export default class Pressability {
-  _config: PressabilityConfig;
-
-  constructor(config: PressabilityConfig) {
-    this._config = config;
-  }
-}