+++ instrumentor/test/fixtures/is-touchable-0-62/code.js
@@ -1,0 +1,1 @@
+class TouchableOpacity extends React.Component<Props, State> {}