+++ js/autotrack/__tests__/common.spec.js
@@ -6,7 +6,7 @@ import { Text, View } from 'react-native';
 const foo = require('react-native');
 import { shallow, mount, render } from 'enzyme';
 
-import { getBaseComponentPropsFromComponent } from '../common';
+import { getBaseComponentProps } from '../common';
 import {
   HeapIgnore,
   HeapIgnoreTargetText,
@@ -53,11 +53,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         target_text: 'foobar',
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@Text;[testID=targetElement];|',
@@ -80,11 +79,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         target_text: 'foobar',
         rn_hierarchy:
           '@WrapperComponent;|@MySpecialComponent;|@Text;[testID=targetElement];|',
@@ -104,11 +102,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         target_text: 'foobar',
         rn_hierarchy:
           '@WrapperComponent;|@ListItem;[rightTitle=React.element];|@Text;[testID=targetElement];|',
@@ -129,9 +126,7 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toBe(null);
     });
 
@@ -146,11 +141,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         target_text: 'foobar',
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@HeapIgnore;|',
@@ -173,11 +167,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         target_text: 'foobar',
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@HeapIgnore;|@Text;|',
@@ -200,11 +193,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@HeapIgnore;|@Text;[testID=targetElement];|',
         source_version: SDK_VERSION,
@@ -229,11 +221,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         target_text: 'foobar',
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@HeapIgnore;|@Text;[testID=targetElement];|',
@@ -265,11 +256,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@HeapIgnore;|@BarFunction;|@HeapIgnore;|',
         source_version: SDK_VERSION,
@@ -289,11 +279,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@withHeapIgnore(Text);[testID=targetElement];|@HeapIgnore;|@Text;|',
         source_version: SDK_VERSION,
@@ -311,11 +300,10 @@ describe('Common autotrack utils', () => {
       const normalComponent = wrapper
         .find({ testID: 'targetElement' })
         .filter(Text);
-      const normalProps = getBaseComponentPropsFromComponent(
-        normalComponent.instance()
-      );
+      const normalProps = getBaseComponentProps(normalComponent.instance());
       expect(normalProps).toEqual({
         is_using_react_navigation_hoc: false,
+        react_native_version: null,
         rn_hierarchy:
           '@WrapperComponent;|@Foo;|@BarClass;|@BarFunction;|@HeapIgnoreTargetText;|@HeapIgnore;|@Text;[testID=targetElement];|',
         source_version: SDK_VERSION,